Add support for arrays
Added support for arrays in ini.decode

Arrays can be defined by suffixing them with two brackets (`[]`).

  foo[] = bar
  foo[] = baz

Foo will be decoded as an array with `bar` and `baz` in the array. There
is a safeguard for accidentally unsetting arrays:

  foo[] = bar
  foo   = baz

This will still be decoded as an array with `bar` and `baz` in the
array. Only when the key is not an array already will the value be
overwritten.

Added array support for ini.encode

`ini.encode` will now output arrays with a pair of brackets (`[]`) as
suffix:

  { ar: [ "one", "two", "three" ] }

will be outputted like as:

  ar[] = one
  ar[] = two
  ar[] = three