Mocha Framework conflict
Running Mocha along with ini module, a conflict occurs when the decode function is called. The conflict is about the input buffer which is not converted to a string explicitly. This fix is required in order to carry on with testing. It is a change in the initialization of the function's data, not in its logic.